backend/requirements.txt
flask
flask-cors
openai-whisper
sqlalchemy
requests
pillow
python-dotenv
